+++ leaves/text_input.mli
@@ -7,53 +7,58 @@ val make :
   ?prompt:string ->
   unit ->
   t
-(**
-   Create a new {Text_input}.
+(** [make text ()] Create a new {Text_input.t}.
+
+    If unspecified, it will use the `default_text_style`, `default_cursor`, and
+    `default_prompt`.
    
-   {[
-     let text_input = Text_input.make "Hello" ()
-   ]}
+    {@ocaml[
+      let text_input = Text_input.make "Hello" ()
+    ]}
 *)
 
 val empty : unit -> t
-(**
-  Create a new {Text_input} with an empty string and default configuration.
+(** [empty ()] creates a new {Text_input.t} with an empty string and default
+    configuration.
 
-  {[
-    let text_input = Text_input.empty ()
-  ]}
+    {@ocaml[
+      let text_input = Text_input.empty ()
+    ]}
 *)
 
 val view : t -> string
-(** 
-    Render the {Text_input} as a string.
+(** [view text_input] renders [text_input] into a string.
     
-    {[
-      let text_input = Text_input.make ~value:"Hello" ()
-      let text = Text_input.view text_input
+    {@ocaml[
+      let text_input = Text_input.make "Hello" () in
+      let text: string = Text_input.view text_input
     ]}
 *)
 
 val update : t -> Minttea.Event.t -> t
-(**
-   Given a {Text_input} and a {Minttea.Event.t} update the {Text_input} to
-   handle cursor movement, text insertions etc.
+(** [update text_input event] updates the [text_input] to handle cursor
+    movement, text insertions etc.
 
-   {[
-     let text_input = Text_input.update text_input event
-   ]}
+    {@ocaml[
+      let text_input = Text_input.update text_input event
+    ]}
 *)
 
 val current_text : t -> string
-(**
-   Get the current text of the {Text_input}.
+(** [current_text t] gets the current text of the {Text_input} [t].
 
-   {[
-     let text = Text_input.current_text text_input
-   ]}
+    {@ocaml[
+      let text = Text_input.current_text text_input
+    ]}
 *)
 
 val set_text : string -> t -> t
-(**
-   Set the current text in the {Text_input}.
+(** [set_text text t] updates the current text of [t] to [text].
+
+    {@ocaml[
+      let text = Text_input.make "hello" () in
+      (* Text_input.current_text text == "hello" *)
+      let new_text = Text_input.set_text "world" text in
+      (* Text_input.current_text new_text == "world" *)
+    ]}
 *)